test: cover a destination the deployment skipped
An AWS::Lambda::EventInvokeConfig sending failures to a Resource simulated
CloudFormation has no implementation for is the case issue 823 was about, and it
now has a test of its own. The template helpers behind it moved to a fixture
under test/, which is what keeps both suites under the file length limit.
